feat: add setting-profile-navbar in profile layout
