+++ components/setting/setting.constants.ts
@@ -1,0 +1,4 @@
+export const SETTING_PROFILE_OPTIONS = [
+  { value: 'about', label: '내 소개' },
+  { value: 'discord', label: '디스코드 프로필' },
+];
